Minimally connect the content component for Categories
Resolves #312
